fix airbrush not showing color picker dialog when doing sneak + right-click on a node which has a formspec in its metadata.
Right-clicking any node without a formspec in its meta will
trigger the color picker as usual, without holding sneak.

This also makes the "use this thing's color" feature use aux1
+ right-click on most nodes, when it was sneak + right-click
before.

On a node with a formspec in its meta, the engine will
intercept the right-click before UD can see it, but then drop
back to UD if the user is holding sneak, so to pick the color
of one of these nodes, use sneak + aux1 + right-click.

(on my setup, sneak is bound to "Shift", aux1 is "E").